requirement_glucose.txt
# Core data processing
pandas>=2.0
numpy>=1.24
nbformat>=5.9

# Notebook / local env support
python-dotenv>=1.0

# Search tool
tavily-python>=0.5.0

# Google Gemini SDK (new client)
google-genai>=1.0.0

# Legacy Gemini SDK (kept only if some notebook cells still import google.generativeai)
google-generativeai>=0.8.0

# The agent framework used by this project
# Install with:
# pip install -U --force-reinstall "git+https://github.com/google-deepmind/onetwo"
